refactor(desktop): split home launch flow
